Instructor Review (VERY LATE)
I am so sorry that I have taken so long to review this! Overall, this is a great application! I hope that you have kept working in JavaScript and that you enjoyed your time in the class. I would refactor some of this code out into separate functions and calls, but overall this is a GREAT project!